requirements.txt
beautifulsoup4==4.8.1
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==1.0.1
django-heroku==0.3.1
djangorestframework==3.10.3
Pillow==6.2.1
psycopg2==2.8.4
PyMySQL==0.9.3
pytz==2019.3
soupsieve==1.9.5
whitenoise==4.1.4
